# [Mono.Android] avoid capturing variables in JNINativeWrapper