--- a/src/Mono.Android/Android.Runtime/JNINativeWrapper.g.cs
+++ b/src/Mono.Android/Android.Runtime/JNINativeWrapper.g.cs
@@ -15,256 +15,256 @@ namespace Android.Runtime
 			return true;
 		}
 
+		internal static void Wrap_JniMarshal_PP_V (this _JniMarshal_PP_V callback, IntPtr jnienv, IntPtr klazz)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
+		internal static void Wrap_JniMarshal_PPI_V (this _JniMarshal_PPI_V callback, IntPtr jnienv, IntPtr klazz, int p0)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz, p0);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
+		internal static IntPtr Wrap_JniMarshal_PPL_L (this _JniMarshal_PPL_L callback, IntPtr jnienv, IntPtr klazz, IntPtr p0)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				return callback (jnienv, klazz, p0);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				return default;
+			}
+		}
+
+		internal static void Wrap_JniMarshal_PPL_V (this _JniMarshal_PPL_V callback, IntPtr jnienv, IntPtr klazz, IntPtr p0)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz, p0);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
+		internal static bool Wrap_JniMarshal_PPL_Z (this _JniMarshal_PPL_Z callback, IntPtr jnienv, IntPtr klazz, IntPtr p0)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				return callback (jnienv, klazz, p0);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				return default;
+			}
+		}
+
+		internal static void Wrap_JniMarshal_PPII_V (this _JniMarshal_PPII_V callback, IntPtr jnienv, IntPtr klazz, int p0, int p1)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz, p0, p1);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
+		internal static void Wrap_JniMarshal_PPLI_V (this _JniMarshal_PPLI_V callback, IntPtr jnienv, IntPtr klazz, IntPtr p0, int p1)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz, p0, p1);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
+		internal static void Wrap_JniMarshal_PPLL_V (this _JniMarshal_PPLL_V callback, IntPtr jnienv, IntPtr klazz, IntPtr p0, IntPtr p1)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz, p0, p1);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
+		internal static bool Wrap_JniMarshal_PPLL_Z (this _JniMarshal_PPLL_Z callback, IntPtr jnienv, IntPtr klazz, IntPtr p0, IntPtr p1)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				return callback (jnienv, klazz, p0, p1);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				return default;
+			}
+		}
+
+		internal static void Wrap_JniMarshal_PPIIL_V (this _JniMarshal_PPIIL_V callback, IntPtr jnienv, IntPtr klazz, int p0, int p1, IntPtr p2)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz, p0, p1, p2);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
+		internal static void Wrap_JniMarshal_PPILL_V (this _JniMarshal_PPILL_V callback, IntPtr jnienv, IntPtr klazz, int p0, IntPtr p1, IntPtr p2)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz, p0, p1, p2);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
+		internal static bool Wrap_JniMarshal_PPLIL_Z (this _JniMarshal_PPLIL_Z callback, IntPtr jnienv, IntPtr klazz, IntPtr p0, int p1, IntPtr p2)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				return callback (jnienv, klazz, p0, p1, p2);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				return default;
+			}
+		}
+
+		internal static IntPtr Wrap_JniMarshal_PPLLL_L (this _JniMarshal_PPLLL_L callback, IntPtr jnienv, IntPtr klazz, IntPtr p0, IntPtr p1, IntPtr p2)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				return callback (jnienv, klazz, p0, p1, p2);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				return default;
+			}
+		}
+
+		internal static bool Wrap_JniMarshal_PPLLL_Z (this _JniMarshal_PPLLL_Z callback, IntPtr jnienv, IntPtr klazz, IntPtr p0, IntPtr p1, IntPtr p2)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				return callback (jnienv, klazz, p0, p1, p2);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				return default;
+			}
+		}
+
+		internal static void Wrap_JniMarshal_PPIIII_V (this _JniMarshal_PPIIII_V callback, IntPtr jnienv, IntPtr klazz, int p0, int p1, int p2, int p3)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz, p0, p1, p2, p3);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
+		internal static void Wrap_JniMarshal_PPLLLL_V (this _JniMarshal_PPLLLL_V callback, IntPtr jnienv, IntPtr klazz, IntPtr p0, IntPtr p1, IntPtr p2, IntPtr p3)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz, p0, p1, p2, p3);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
+		internal static void Wrap_JniMarshal_PPLIIII_V (this _JniMarshal_PPLIIII_V callback, IntPtr jnienv, IntPtr klazz, IntPtr p0, int p1, int p2, int p3, int p4)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz, p0, p1, p2, p3, p4);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
+		internal static void Wrap_JniMarshal_PPZIIII_V (this _JniMarshal_PPZIIII_V callback, IntPtr jnienv, IntPtr klazz, bool p0, int p1, int p2, int p3, int p4)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz, p0, p1, p2, p3, p4);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
+		internal static void Wrap_JniMarshal_PPLIIIIIIII_V (this _JniMarshal_PPLIIIIIIII_V callback, IntPtr jnienv, IntPtr klazz, IntPtr p0, int p1, int p2, int p3, int p4, int p5, int p6, int p7, int p8)
+		{
+			JNIEnv.WaitForBridgeProcessing ();
+			try {
+				callback (jnienv, klazz, p0, p1, p2, p3, p4, p5, p6, p7, p8);
+			} catch (Exception e) when (_unhandled_exception (e)) {
+				AndroidEnvironment.UnhandledException (e);
+				
+			}
+		}
+
 		private static Delegate CreateBuiltInDelegate (Delegate dlg, Type delegateType)
 		{
 			switch (delegateType.Name) {
-				case nameof (_JniMarshal_PP_V): {
-					_JniMarshal_PP_V callback = Unsafe.As<_JniMarshal_PP_V> (dlg);
-					_JniMarshal_PP_V result = (jnienv, klazz) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPI_V): {
-					_JniMarshal_PPI_V callback = Unsafe.As<_JniMarshal_PPI_V> (dlg);
-					_JniMarshal_PPI_V result = (jnienv, klazz, p0) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz, p0);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPL_L): {
-					_JniMarshal_PPL_L callback = Unsafe.As<_JniMarshal_PPL_L> (dlg);
-					_JniMarshal_PPL_L result = (jnienv, klazz, p0) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							return callback (jnienv, klazz, p0);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							return default;
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPL_V): {
-					_JniMarshal_PPL_V callback = Unsafe.As<_JniMarshal_PPL_V> (dlg);
-					_JniMarshal_PPL_V result = (jnienv, klazz, p0) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz, p0);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPL_Z): {
-					_JniMarshal_PPL_Z callback = Unsafe.As<_JniMarshal_PPL_Z> (dlg);
-					_JniMarshal_PPL_Z result = (jnienv, klazz, p0) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							return callback (jnienv, klazz, p0);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							return default;
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPII_V): {
-					_JniMarshal_PPII_V callback = Unsafe.As<_JniMarshal_PPII_V> (dlg);
-					_JniMarshal_PPII_V result = (jnienv, klazz, p0, p1) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz, p0, p1);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPLI_V): {
-					_JniMarshal_PPLI_V callback = Unsafe.As<_JniMarshal_PPLI_V> (dlg);
-					_JniMarshal_PPLI_V result = (jnienv, klazz, p0, p1) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz, p0, p1);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPLL_V): {
-					_JniMarshal_PPLL_V callback = Unsafe.As<_JniMarshal_PPLL_V> (dlg);
-					_JniMarshal_PPLL_V result = (jnienv, klazz, p0, p1) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz, p0, p1);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPLL_Z): {
-					_JniMarshal_PPLL_Z callback = Unsafe.As<_JniMarshal_PPLL_Z> (dlg);
-					_JniMarshal_PPLL_Z result = (jnienv, klazz, p0, p1) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							return callback (jnienv, klazz, p0, p1);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							return default;
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPIIL_V): {
-					_JniMarshal_PPIIL_V callback = Unsafe.As<_JniMarshal_PPIIL_V> (dlg);
-					_JniMarshal_PPIIL_V result = (jnienv, klazz, p0, p1, p2) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz, p0, p1, p2);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPILL_V): {
-					_JniMarshal_PPILL_V callback = Unsafe.As<_JniMarshal_PPILL_V> (dlg);
-					_JniMarshal_PPILL_V result = (jnienv, klazz, p0, p1, p2) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz, p0, p1, p2);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPLIL_Z): {
-					_JniMarshal_PPLIL_Z callback = Unsafe.As<_JniMarshal_PPLIL_Z> (dlg);
-					_JniMarshal_PPLIL_Z result = (jnienv, klazz, p0, p1, p2) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							return callback (jnienv, klazz, p0, p1, p2);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							return default;
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPLLL_L): {
-					_JniMarshal_PPLLL_L callback = Unsafe.As<_JniMarshal_PPLLL_L> (dlg);
-					_JniMarshal_PPLLL_L result = (jnienv, klazz, p0, p1, p2) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							return callback (jnienv, klazz, p0, p1, p2);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							return default;
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPLLL_Z): {
-					_JniMarshal_PPLLL_Z callback = Unsafe.As<_JniMarshal_PPLLL_Z> (dlg);
-					_JniMarshal_PPLLL_Z result = (jnienv, klazz, p0, p1, p2) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							return callback (jnienv, klazz, p0, p1, p2);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							return default;
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPIIII_V): {
-					_JniMarshal_PPIIII_V callback = Unsafe.As<_JniMarshal_PPIIII_V> (dlg);
-					_JniMarshal_PPIIII_V result = (jnienv, klazz, p0, p1, p2, p3) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz, p0, p1, p2, p3);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPLLLL_V): {
-					_JniMarshal_PPLLLL_V callback = Unsafe.As<_JniMarshal_PPLLLL_V> (dlg);
-					_JniMarshal_PPLLLL_V result = (jnienv, klazz, p0, p1, p2, p3) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz, p0, p1, p2, p3);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPLIIII_V): {
-					_JniMarshal_PPLIIII_V callback = Unsafe.As<_JniMarshal_PPLIIII_V> (dlg);
-					_JniMarshal_PPLIIII_V result = (jnienv, klazz, p0, p1, p2, p3, p4) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz, p0, p1, p2, p3, p4);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPZIIII_V): {
-					_JniMarshal_PPZIIII_V callback = Unsafe.As<_JniMarshal_PPZIIII_V> (dlg);
-					_JniMarshal_PPZIIII_V result = (jnienv, klazz, p0, p1, p2, p3, p4) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz, p0, p1, p2, p3, p4);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
-				case nameof (_JniMarshal_PPLIIIIIIII_V): {
-					_JniMarshal_PPLIIIIIIII_V callback = Unsafe.As<_JniMarshal_PPLIIIIIIII_V> (dlg);
-					_JniMarshal_PPLIIIIIIII_V result = (jnienv, klazz, p0, p1, p2, p3, p4, p5, p6, p7, p8) => {
-						JNIEnv.WaitForBridgeProcessing ();
-						try {
-							callback (jnienv, klazz, p0, p1, p2, p3, p4, p5, p6, p7, p8);
-						} catch (Exception e) when (_unhandled_exception (e)) {
-							AndroidEnvironment.UnhandledException (e);
-							
-						}
-					};
-					return result;
-				}
+				case nameof (_JniMarshal_PP_V):
+					return new _JniMarshal_PP_V (Unsafe.As<_JniMarshal_PP_V> (dlg).Wrap_JniMarshal_PP_V);
+				case nameof (_JniMarshal_PPI_V):
+					return new _JniMarshal_PPI_V (Unsafe.As<_JniMarshal_PPI_V> (dlg).Wrap_JniMarshal_PPI_V);
+				case nameof (_JniMarshal_PPL_L):
+					return new _JniMarshal_PPL_L (Unsafe.As<_JniMarshal_PPL_L> (dlg).Wrap_JniMarshal_PPL_L);
+				case nameof (_JniMarshal_PPL_V):
+					return new _JniMarshal_PPL_V (Unsafe.As<_JniMarshal_PPL_V> (dlg).Wrap_JniMarshal_PPL_V);
+				case nameof (_JniMarshal_PPL_Z):
+					return new _JniMarshal_PPL_Z (Unsafe.As<_JniMarshal_PPL_Z> (dlg).Wrap_JniMarshal_PPL_Z);
+				case nameof (_JniMarshal_PPII_V):
+					return new _JniMarshal_PPII_V (Unsafe.As<_JniMarshal_PPII_V> (dlg).Wrap_JniMarshal_PPII_V);
+				case nameof (_JniMarshal_PPLI_V):
+					return new _JniMarshal_PPLI_V (Unsafe.As<_JniMarshal_PPLI_V> (dlg).Wrap_JniMarshal_PPLI_V);
+				case nameof (_JniMarshal_PPLL_V):
+					return new _JniMarshal_PPLL_V (Unsafe.As<_JniMarshal_PPLL_V> (dlg).Wrap_JniMarshal_PPLL_V);
+				case nameof (_JniMarshal_PPLL_Z):
+					return new _JniMarshal_PPLL_Z (Unsafe.As<_JniMarshal_PPLL_Z> (dlg).Wrap_JniMarshal_PPLL_Z);
+				case nameof (_JniMarshal_PPIIL_V):
+					return new _JniMarshal_PPIIL_V (Unsafe.As<_JniMarshal_PPIIL_V> (dlg).Wrap_JniMarshal_PPIIL_V);
+				case nameof (_JniMarshal_PPILL_V):
+					return new _JniMarshal_PPILL_V (Unsafe.As<_JniMarshal_PPILL_V> (dlg).Wrap_JniMarshal_PPILL_V);
+				case nameof (_JniMarshal_PPLIL_Z):
+					return new _JniMarshal_PPLIL_Z (Unsafe.As<_JniMarshal_PPLIL_Z> (dlg).Wrap_JniMarshal_PPLIL_Z);
+				case nameof (_JniMarshal_PPLLL_L):
+					return new _JniMarshal_PPLLL_L (Unsafe.As<_JniMarshal_PPLLL_L> (dlg).Wrap_JniMarshal_PPLLL_L);
+				case nameof (_JniMarshal_PPLLL_Z):
+					return new _JniMarshal_PPLLL_Z (Unsafe.As<_JniMarshal_PPLLL_Z> (dlg).Wrap_JniMarshal_PPLLL_Z);
+				case nameof (_JniMarshal_PPIIII_V):
+					return new _JniMarshal_PPIIII_V (Unsafe.As<_JniMarshal_PPIIII_V> (dlg).Wrap_JniMarshal_PPIIII_V);
+				case nameof (_JniMarshal_PPLLLL_V):
+					return new _JniMarshal_PPLLLL_V (Unsafe.As<_JniMarshal_PPLLLL_V> (dlg).Wrap_JniMarshal_PPLLLL_V);
+				case nameof (_JniMarshal_PPLIIII_V):
+					return new _JniMarshal_PPLIIII_V (Unsafe.As<_JniMarshal_PPLIIII_V> (dlg).Wrap_JniMarshal_PPLIIII_V);
+				case nameof (_JniMarshal_PPZIIII_V):
+					return new _JniMarshal_PPZIIII_V (Unsafe.As<_JniMarshal_PPZIIII_V> (dlg).Wrap_JniMarshal_PPZIIII_V);
+				case nameof (_JniMarshal_PPLIIIIIIII_V):
+					return new _JniMarshal_PPLIIIIIIII_V (Unsafe.As<_JniMarshal_PPLIIIIIIII_V> (dlg).Wrap_JniMarshal_PPLIIIIIIII_V);
 				default:
 					return null;
 			}