One performance issue with the existing code:

    _JniMarshal_PP_V callback = Unsafe.As<_JniMarshal_PP_V> (dlg);
    _JniMarshal_PP_V result = (jnienv, klazz) => {
        JNIEnv.WaitForBridgeProcessing ();
        try {
            callback (jnienv, klazz);
        } catch (Exception e) when (_unhandled_exception (e)) {
            AndroidEnvironment.UnhandledException (e);
        }
    };
    return result;

In this example, `callback` is "captured", and an intermediate class
is generated to store this data such as
`Android.Runtime.JNINativeWrapper/<>c__DisplayClass5_15`.

A way to do this is:

    return new _JniMarshal_PP_V (Unsafe.As<_JniMarshal_PP_V> (dlg).Wrap_JniMarshal_PP_V);

Where `Wrap_JniMarshal_PP_V` is an extension method:

    internal static void Wrap_JniMarshal_PP_V (this _JniMarshal_PP_V callback, IntPtr jnienv, IntPtr klazz)
    {
        JNIEnv.WaitForBridgeProcessing ();
        try {
            callback (jnienv, klazz);
        } catch (Exception e) when (_unhandled_exception (e)) {
            AndroidEnvironment.UnhandledException (e);
        }
    }

This avoids the captured variable *and* uses "regular" methods.

## Results ##

Running a regular (non-AOT) Release build of `dotnet new maui` on a
Pixel 5:

    Before:
    02-07 14:47:25.935  1566  1566 I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::493336
    02-07 14:47:27.158  1635  1635 I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::234846
    02-07 14:47:28.347  1737  1737 I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::245159
    02-07 14:47:29.562  1794  1794 I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::511304
    02-07 14:47:30.732  1850  1850 I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::256304
    02-07 14:47:31.913  1956  1956 I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::323647
    02-07 14:47:33.143  2043  2043 I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::58491
    02-07 14:47:34.355  2137  2137 I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::172086
    02-07 14:47:35.559  2217  2217 I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::335627
    02-07 14:47:36.789  2352  2352 I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::679325
    Average(ms): 19.3310125
    Std Err(ms): 0.0579249384982648
    Std Dev(ms): 0.18317473897969
    After:
    02-07 14:42:24.153 30712 30712 I monodroid-timing: Runtime.register: end time; elapsed: 0s:18::694950
    02-07 14:42:25.353 30770 30770 I monodroid-timing: Runtime.register: end time; elapsed: 0s:18::700315
    02-07 14:42:26.535 30825 30825 I monodroid-timing: Runtime.register: end time; elapsed: 0s:18::764169
    02-07 14:42:27.733 30887 30887 I monodroid-timing: Runtime.register: end time; elapsed: 0s:18::940054
    02-07 14:42:28.984 30950 30950 I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::255731
    02-07 14:42:30.161 31013 31013 I monodroid-timing: Runtime.register: end time; elapsed: 0s:18::712554
    02-07 14:42:31.367 31070 31070 I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::198700
    02-07 14:42:32.555 31126 31126 I monodroid-timing: Runtime.register: end time; elapsed: 0s:18::775367
    02-07 14:42:33.810 31190 31190 I monodroid-timing: Runtime.register: end time; elapsed: 0s:18::937189
    02-07 14:42:34.997 31245 31245 I monodroid-timing: Runtime.register: end time; elapsed: 0s:18::897606
    Average(ms): 18.8876635
    Std Err(ms): 0.0639684356019523
    Std Dev(ms): 0.202285954859973

This was using:

    msbuild Xamarin.Android.sln -t:InstallMaui -p:MauiVersion=6.0.200-preview.14.2816

After this change goes in, I will update the built-in AOT profiles at
a later date.