@jonpryor 2022-02-10
Draft commit message:

```markdown
Context: 32cff4383232d5de156bd6c5d10292fcffa66d50

One performance issue with the existing code from 32cff438:

	_JniMarshal_PP_V callback = (_JniMarshal_PP_V) Delegate.CreateDelegate (typeof (_JniMarshal_PP_V), dlg.Target, dlg.Method);
	_JniMarshal_PP_V result = (jnienv, klazz) => {
	    JNIEnv.WaitForBridgeProcessing ();
	    try {
	        callback (jnienv, klazz);
	    } catch (Exception e) when (_unhandled_exception (e)) {
	        AndroidEnvironment.UnhandledException (e);
	    }
	};
	return result;

is that `callback` is "captured", and an intermediate class is
generated by the compiler to store this data, such as
`Android.Runtime.JNINativeWrapper/<>c__DisplayClass5_15`.

The problem with this generated type is that:

 1. It's additional overhead, but more importantly
 2. Profiled AOT does not appear to be able to AOT this method.

In regards to (2), `adb logcat` contains:

	D Mono : AOT NOT FOUND: (wrapper native-to-managed) Android.Runtime.JNINativeWrapper/<>c__DisplayClass5_15:<CreateBuiltInDelegate>b__15 (intptr,intptr,intptr,intptr,intptr,intptr).
	D Mono : AOT: FOUND method Android.Runtime.JNINativeWrapper/<>c__DisplayClass5_15:<CreateBuiltInDelegate>b__15 (intptr,intptr,intptr,intptr,intptr,intptr) [0x732a050760 - 0x732a0508d0 0x732a05cb37]

Because of this issue, if I update the built-in AOT profile -- it
seems to make startup performance *worse*!

One instance of the method is a "trampoline" that always uses the
JIT, so we'd be better off using "regular" methods instead of
anonymous methods here.

A way to do this is:

	return new _JniMarshal_PP_V (Unsafe.As<_JniMarshal_PP_V> (dlg).Wrap_JniMarshal_PP_V);

Where `Wrap_JniMarshal_PP_V` is an extension method:

	internal static void Wrap_JniMarshal_PP_V (this _JniMarshal_PP_V callback, IntPtr jnienv, IntPtr klazz)
	{
	    JNIEnv.WaitForBridgeProcessing ();
	    try {
	        callback (jnienv, klazz);
	    } catch (Exception e) when (_unhandled_exception (e)) {
	        AndroidEnvironment.UnhandledException (e);
	    }
	}

This avoids the captured variable *and* uses "regular" methods.

~~ Results ~~

Running a regular (non-AOT) Release build of `dotnet new maui` on a
Pixel 5:

  * Before:

        I monodroid-timing: Runtime.register: end time; elapsed: 0s:19::493336
        …

      * Average(ms): 19.3310125
      * Std Err(ms): 0.0579249384982648
      * Std Dev(ms): 0.18317473897969

  * After:

        I monodroid-timing: Runtime.register: end time; elapsed: 0s:18::694950
        …

      * Average(ms): 18.8876635
      * Std Err(ms): 0.0639684356019523
      * Std Dev(ms): 0.202285954859973

This was using:

	msbuild Xamarin.Android.sln -t:InstallMaui -p:MauiVersion=6.0.200-preview.14.2816

After this change goes in, I will update the built-in AOT profiles at
a later date.
```