@jonpryor 2022-02-10
@jonathanpeppers: one thing about the [commit message](https://github.com/xamarin/xamarin-android/pull/6713#issuecomment-1034440269) confuses me; you originally wrote:

> The problem here is that this is overhead and Profiled AOT does not
appear to be able to AOT this method:
> 
>     02-03 17:21:42.003 5354 5354 D Mono : AOT NOT FOUND: (wrapper native-to-managed) Android.Runtime.JNINativeWrapper/<>c__DisplayClass5_15:<CreateBuiltInDelegate>b__15 (intptr,intptr,intptr,intptr,intptr,intptr).
>     02-03 17:21:42.003 5354 5354 D Mono : AOT: FOUND method Android.Runtime.JNINativeWrapper/<>c__DisplayClass5_15:<CreateBuiltInDelegate>b__15 (intptr,intptr,intptr,intptr,intptr,intptr) [0x732a050760 - 0x732a0508d0 0x732a05cb37]

Yet, as you later note, the `(wrapper native-to-managed)` methods are *always* JIT'd ("One instance of the method is a "trampoline" that always uses the JIT"), and the other method *is* found in the AOT cache!

Thus, *yay*, this improves the average by 0.5ms, but the *explanation* feels somewhat lacking?

What am I missing?