@jonathanpeppers 2022-02-10
@jonpryor  I updated your commit message above: https://github.com/xamarin/xamarin-android/pull/6713#issuecomment-1034440269

```
After recording an AOT profile with these changes, we get messages like:

    02-09 17:16:13.205 26543 26543 D Mono    : AOT: FOUND method Android.Runtime.JNINativeWrapper:Wrap_JniMarshal_PP_V (_JniMarshal_PP_V,intptr,intptr) [0x731d781930 - 0x731d781a70 0x731d7a4ead]
```

I recorded a profile for the podcast app, and I'm seeing this show up now.